requirements.txt
flask
flask_login
flask_socketio
requests
pandas
pyserial
rsa
openpyxl